requirements.txt
# Core dependencies
#torch>=2.4.1
#torchvision>=0.15.0
transformers==4.45.0
diffusers==0.30.0
accelerate==0.33.0

# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# Utilities
protobuf==4.23.4
sentencepiece==0.2.0
